+++ .gulp/images.js
@@ -1,23 +1,17 @@
 import gulp from 'gulp';
-const { src, dest } = gulp
+const { src, dest } = gulp;
 
-import cache from 'gulp-cache'
+import cache from 'gulp-cache';
 
 /**
  * Images
  */
-import imagemin from 'gulp-imagemin'
-import imageminPngquant from 'imagemin-pngquant'
-import imageminZopfli from 'imagemin-zopfli'
-import imageminMozjpeg from 'imagemin-mozjpeg'
-import imageminJpegRecompress from 'imagemin-jpeg-recompress'
-import imageminGiflossy from 'imagemin-giflossy'
-import webp from 'gulp-webp'
+import webp from 'gulp-webp';
 
 /**
  * SVG
  */
-import sprite from 'gulp-svg-sprite'
+import sprite from 'gulp-svg-sprite';
 
 
 /**
@@ -31,66 +25,22 @@ import sprite from 'gulp-svg-sprite'
 const cfg = {
   src: {
     img: './src/images/**/*',
-    webp: './dist/images/**/*.{png,jpg,jpeg}',
+    webp: './src/images/**/*.{png,jpg,jpeg}',
   },
   dest: {
     img: './dist/images/',
-    webp: './dist/images/**/*.{png,jpg,jpeg}',
+    webp: './dist/images/',
   },
   build: {
     img: './build/images/',
   }
-}
+};
 
 /**
  * Images
  * ================================================================================
  */
 
-// Images Minify
-const imagesCompress = () =>
-  src(cfg.src.img)
-    .pipe(
-      cache(
-        imagemin([
-          imageminPngquant({
-            speed: 1,
-            quality: [0.95, 1],
-          }),
-          imageminZopfli({
-            more: true,
-          }),
-          imageminGiflossy({
-            optimizationLevel: 3,
-            optimize: 3, //keep-empty: Preserve empty transparent frames
-            lossy: 2,
-          }),
-          imagemin.svgo({
-            plugins: [
-              { optimizationLevel: 3 },
-              { progressive: true },
-              { interlaced: true },
-              { removeViewBox: false },
-              { removeUselessStrokeAndFill: false },
-              { cleanupIDs: false },
-            ],
-          }),
-          imagemin.mozjpeg({
-            progressive: true,
-          }),
-          imageminJpegRecompress({
-            loops: 6,
-            min: 40,
-            max: 85,
-            quality: 'low',
-          }),
-          imageminMozjpeg({
-            quality: 90,
-          }),
-        ])
-      )
-    )
-    .pipe(dest(cfg.dest.img))
 
 /**
  * Webp
@@ -100,12 +50,10 @@ const webpCompress = () =>
   src(cfg.src.webp)
     .pipe(
       cache(
-        webp({
-          quality: 75,
-        })
+        webp()
       )
     )
-    .pipe(dest(cfg.dest.img))
+    .pipe(dest(cfg.dest.img));
 
 /**
  * SVG Sprite
@@ -128,8 +76,8 @@ const genSvgSprite = () =>
         },
       })
     )
-    .pipe(dest('./src/images/'))
+    .pipe(dest('./src/images/'));
 
-const imgCopy = () => src('./src/images/**/*').pipe(dest('./dist/images/')).pipe(dest('./build/images/'))
+const imgCopy = () => src('./src/images/**/*').pipe(dest('./dist/images/')).pipe(dest('./build/images/'));
 
-export { imagesCompress, webpCompress, genSvgSprite, imgCopy }
+export { webpCompress, genSvgSprite, imgCopy };