+++ .gulp/javascript.js
@@ -1,35 +1,40 @@
 import gulp from 'gulp';
-const { src, dest } = gulp
+const { src, dest } = gulp;
 
 
 /**
  * System
  */
-import path from 'path'
-import connect from 'gulp-connect'
+import path from 'path';
+import connect from 'gulp-connect';
 
 
 /**
  * Notification
  */
-import plumber from 'gulp-plumber'
-import notify from 'gulp-notify'
+import plumber from 'gulp-plumber';
+import notify from 'gulp-notify';
 import size from 'gulp-size';
 
 /**
  * Compressors
  */
-import uglify from 'gulp-uglify'
+import uglify from 'gulp-uglify';
 
 /**
  * JavaScript
  */
 import { rollup } from 'rollup';
-import { babel } from '@rollup/plugin-babel'
-import alias from '@rollup/plugin-alias'
-import commonjs from '@rollup/plugin-commonjs'
-import { nodeResolve } from '@rollup/plugin-node-resolve'
-import standard from 'gulp-standard'
+
+import { babel } from '@rollup/plugin-babel';
+import alias from '@rollup/plugin-alias';
+import commonjs from '@rollup/plugin-commonjs';
+import json from "@rollup/plugin-json";
+import { nodeResolve } from '@rollup/plugin-node-resolve';
+import url from '@rollup/plugin-url';
+import { string } from "rollup-plugin-string";
+
+import standard from 'gulp-standard';
 
 /**
  * Settings
@@ -51,7 +56,7 @@ const cfg = {
     output: './dist/javascript/app.js',
     format: 'iife',
   },
-}
+};
 
 
 const __dirname = path.resolve(path.dirname(''));
@@ -70,11 +75,25 @@ const roll = () =>
     input: cfg.roll.input,
     plugins: [
       babel({
+        exclude: "node_modules/**",
+        presets: ["@babel/preset-env"],
+        // plugins: ["babel-plugin-transform-html-import-to-string"],
+        // extensions: [".js", ".html"],
         babelHelpers: 'bundled',
       }),
+      url(),
+      string({
+        // Required to be specified
+        include: "**/*.html",
+
+        // Undefined by default
+        exclude: ["**/index.html"]
+      }),
+      json(),
       alias({
         entries: [
           { find: 'src', replacement: `${__dirname}/src/javascript/` },
+          { find: 'root', replacement: `${__dirname}/` }
         ],
       }),
       nodeResolve({
@@ -94,12 +113,12 @@ const roll = () =>
         format: cfg.roll.format,
         name: 'library',
         sourcemap: false,
-      })
-    })
+      });
+    });
 
 // Reload Browser after JS Changes
 const scriptsReload = () => src(cfg.src.js)
-  .pipe(connect.reload())
+  .pipe(connect.reload());
 
 
 // JS Minify
@@ -109,7 +128,7 @@ const compressJS = () =>
     .pipe(uglify())
     .on('error', notify.onError())
     .pipe(size())
-    .pipe(dest('./build/javascript/'))
+    .pipe(dest('./build/javascript/'));
 
 
 const standardJS = () =>
@@ -118,7 +137,7 @@ const standardJS = () =>
     .pipe(standard.reporter('default', {
       breakOnError: false,
       quiet: true
-    }))
+    }));
 
 
-export { roll, scriptsReload, compressJS, standardJS }
+export { roll, scriptsReload, compressJS, standardJS };