+++ gulpfile.js
@@ -15,10 +15,10 @@ import { roll, scriptsReload, compressJS } from './.gulp/javascript.js';
 import { htmlGenerate, htmlReload, htmlCompress, validateHtml, testHtml, htmlPagesPreview } from './.gulp/html.js';
 
 // Images
-import { imagesCompress, webpCompress, genSvgSprite, imgCopy } from './.gulp/images.js';
+import { webpCompress, genSvgSprite, imgCopy } from './.gulp/images.js';
 
 // Server
-import { openServer, openBrowser, bumper, clean, cleanDist, lt } from './.gulp/server.js';
+import { openServer, openBrowser, bumper, cleanBuild, cleanDist, cleanHTML, lt } from './.gulp/server.js';
 
 // Misc
 import { copyFiles, copyBuildFiles, copyVideo, copyFonts, copyIcons } from './.gulp/misc.js';
@@ -64,9 +64,9 @@ const jsVendorLibs = () =>
  */
 const watcher = () => {
   watch('./src/scss/**/*.scss', series(scss, stylesReload));
-  watch('./src/**/*.html', series(htmlGenerate, cleanDist, htmlReload, testHtml));
+  watch('./src/**/*.html', series(htmlGenerate, cleanHTML, htmlReload, testHtml));
   watch('./src/javascript/**/*.js', series(series(roll, jsVendorLibs), scriptsReload));
-  watch('./src/images/**/*', imgCopy);
+  watch('./src/images/**/*', series(webpCompress, imgCopy));
   watch('./src/favicons/**/*', copyIcons);
   watch('./src/fonts/**/*', copyFonts);
   watch('./src/video/**/*', copyVideo);
@@ -76,6 +76,8 @@ const watcher = () => {
  * Default Tasks
  */
 export default series(
+  cleanDist,
+  cleanBuild,
   parallel(
     copyFiles,
     copyVideo,
@@ -83,8 +85,8 @@ export default series(
     copyIcons,
     series(series(roll, jsVendorLibs)),
     series(scss),
-    series(imgCopy),
-    series(htmlGenerate, cleanDist, openBrowser, htmlPagesPreview, validateHtml),
+    series(webpCompress, imgCopy),
+    series(htmlGenerate, cleanHTML, openBrowser, htmlPagesPreview, validateHtml),
     openServer,
     lt,
     watcher
@@ -100,21 +102,22 @@ const test = parallel(mobileTestRes, htmlSpeedRes, cssTestRes);
  * Build Tasks
  */
 const build = series(
-  clean,
+  cleanDist,
+  cleanBuild,
   copyBuildFiles,
   parallel(
     series(series(roll, jsVendorLibs), compressJS),
     series(scss, cssCompress),
-    series(imgCopy),
-    series(htmlGenerate, cleanDist, htmlCompress)
+    series(webpCompress, imgCopy),
+    series(htmlGenerate, cleanHTML, htmlCompress)
   ),
   bumper
 );
 
 /**
- * Generate Compressed Images
+ * Generate Images
  */
-const images = series(imagesCompress, webpCompress);
+const images = series(imgCopy, webpCompress);
 
 /**
  * Generate SVG Sprite